Added Bottom Navigation Rev #1
