+++ app/src/main/java/com/taio/taio/MainActivity.kt
@@ -3,16 +3,9 @@ package com.taio.taio
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.taio.taio.ui.TandatanganioApp
-import com.taio.taio.ui.TandatanganioApp
-import com.taio.taio.ui.bottomnav.MainScreen
 import com.taio.taio.ui.theme.TandatanganioMobileTheme
 
 class MainActivity : ComponentActivity() {
@@ -20,7 +13,7 @@ class MainActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             TandatanganioMobileTheme {
-                MainScreen()
+                TandatanganioApp()
             }
         }
     }
