--- app/src/main/java/com/taio/taio/ui/bottomnav/MainScreen.kt
@@ -1,113 +1,0 @@
-package com.taio.taio.ui.bottomnav
-
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Home
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.navigation.NavDestination
-import androidx.navigation.NavDestination.Companion.hierarchy
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.currentBackStackEntryAsState
-import androidx.navigation.compose.rememberNavController
-import com.taio.taio.ui.theme.fonts
-
-@Preview
-@Composable
-fun MainScreen() {
-    val navController = rememberNavController()
-    Scaffold(
-        bottomBar = {
-            BottomBar(navController = navController)
-        },
-        floatingActionButtonPosition = FabPosition.Center,
-        isFloatingActionButtonDocked = true,
-        floatingActionButton = {
-            FloatingActionButton(
-                shape = CircleShape,
-                onClick = { TODO() },
-                contentColor = Color.White
-            ) {
-                Icon(imageVector = Icons.Default.Home, contentDescription = "Add icon")
-            }
-        }
-    ) {
-        BottomNavGraph(navController = navController)
-    }
-}
-
-
-@Composable
-fun BottomBar(navController: NavHostController) {
-    val screens = listOf(
-        BottomBarScreen.Home,
-        BottomBarScreen.Submission,
-//        BottomBarScreen.Generate,
-        BottomBarScreen.Request,
-        BottomBarScreen.Profile
-    )
-    val navBackStackEntry by navController.currentBackStackEntryAsState()
-    val currentDestination = navBackStackEntry?.destination
-
-    BottomNavigation(
-        modifier = Modifier
-            .height(70.dp)
-            .fillMaxWidth(),
-        // backgroundColor = Color.White,
-        elevation = 0.dp
-    ) {
-        screens.forEach { screen ->
-            AddItem(
-                screen = screen,
-                currentDestination = currentDestination,
-                navController = navController
-            )
-        }
-    }
-
-}
-
-@Composable
-fun RowScope.AddItem(
-    screen: BottomBarScreen,
-    currentDestination: NavDestination?,
-    navController: NavHostController
-) {
-    BottomNavigationItem(
-        modifier = Modifier
-            .padding(bottom = 5.dp),
-        icon = {
-            Icon(
-                painter = painterResource(id = screen.icon),
-                contentDescription = "Navigate Icon"
-            )
-        },
-        label = {
-            Text(
-                text = screen.title,
-                fontFamily = fonts,
-                fontWeight = FontWeight.Medium,
-                fontSize = 10.sp
-            )
-        },
-        selected = currentDestination?.hierarchy?.any {
-            it.route == screen.route
-        } == true,
-        selectedContentColor = Color.White,
-        unselectedContentColor = Color.White.copy(alpha = 0.4f),
-        onClick = {
-            navController.navigate(screen.route)
-        }
-    )
-
-}
-
